Move ref to separate file, disable logging by default
